refactor: db into files
