+++ components/ordhook-core/src/core/mod.rs
@@ -12,17 +12,17 @@ use chainhook_sdk::utils::Context;
 
 use crate::{
     config::{Config, LogConfig, MetaProtocolsConfig, ResourcesConfig},
-    db::blocks::{
-        find_last_block_inserted, find_pinned_block_bytes_at_block_height,
-        open_ordhook_db_conn_rocks_db_loop,
+    db::{
+        blocks::{
+            find_last_block_inserted, find_pinned_block_bytes_at_block_height,
+            open_ordhook_db_conn_rocks_db_loop,
+        },
+        cursor::TransactionBytesCursor,
+        ordinals::{find_latest_inscription_block_height, open_readonly_ordhook_db_conn},
     },
     initialize_databases,
     utils::bitcoind::bitcoind_get_block_height,
 };
-
-use crate::db::{find_latest_inscription_block_height, open_readonly_ordhook_db_conn};
-
-use crate::db::TransactionBytesCursor;
 
 #[derive(Clone, Debug)]
 pub struct OrdhookConfig {
