+++ components/ordhook-core/src/service/mod.rs
@@ -25,13 +25,13 @@ use crate::core::{new_traversals_lazy_cache, should_sync_ordhook_db, should_sync
 use crate::db::blocks::{
     find_missing_blocks, insert_entry_in_blocks, open_ordhook_db_conn_rocks_db_loop, run_compaction,
 };
-use crate::db::update_sequence_metadata_with_block;
-use crate::db::{
-    delete_data_in_ordhook_db, find_latest_inscription_block_height,
-    get_latest_indexed_inscription_number, open_readonly_ordhook_db_conn,
-    open_readwrite_ordhook_dbs, update_ordinals_db_with_block, BlockBytesCursor,
-    TransactionBytesCursor,
+use crate::db::cursor::{BlockBytesCursor, TransactionBytesCursor};
+use crate::db::ordinals::{
+    find_latest_inscription_block_height, get_latest_indexed_inscription_number,
+    open_readonly_ordhook_db_conn, update_ordinals_db_with_block,
+    update_sequence_metadata_with_block,
 };
+use crate::db::{delete_data_in_ordhook_db, open_readwrite_ordhook_dbs};
 use crate::scan::bitcoin::process_block_with_predicates;
 use crate::service::observers::create_and_consolidate_chainhook_config_with_predicates;
 use crate::service::runloops::start_bitcoin_scan_runloop;
@@ -636,7 +636,6 @@ fn chainhook_sidecar_mutate_ordhook_db(command: HandleBlock, config: &Config, ct
             }
 
             update_ordinals_db_with_block(&block, &inscriptions_db_conn_rw, ctx);
-
             update_sequence_metadata_with_block(&block, &inscriptions_db_conn_rw, &ctx);
 
             if let Some(brc20_conn_rw) = brc20_db_conn_rw {
